split files into upload, download
