+++ crates/api_crud/src/site/read.rs
@@ -69,6 +69,6 @@ async fn read_site(context: &LemmyContext) -> LemmyResult<GetSiteResponse> {
     tagline,
     oauth_providers: Some(oauth_providers),
     admin_oauth_providers: Some(admin_oauth_providers),
-    image_upload_disabled: context.settings().pictrs()?.disable_image_upload,
+    image_upload_disabled: context.settings().pictrs()?.image_upload_disabled,
   })
 }
